rename: StyledText, StyledButton 컴포넌트 디렉토리 이동 및 경로 업데이트
- ref: #22
