+++ src/components/common/StyledText/index.tsx
