+++ src/screens/SocialLogin/index.tsx
@@ -6,8 +6,8 @@ import { StackNavigationProp } from '@react-navigation/stack';
 import styled from 'styled-components/native';
 import { StackParamList } from '@type/ScreenParamList';
 import responsive from '@utils/responsive';
-import StyledText from '@styles/StyledText';
-import StyledButton from '@styles/StyledButton';
+import StyledText from '@components/common/StyledText';
+import StyledButton from '@components/common/StyledButton';
 import ScreenLayout from '@screens/ScreenLayout';
 import SocialLoginWebView from '@screens/SocialLogin/SocialLoginWebView';
 import KakaoIcon from '@assets/icons/kakao.svg';
